discord_bot/requirements.txt
discord.py
dotenv
flask
yt-dlp